requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.5
django-admin-sortable==2.2.3
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.1
gunicorn==20.0.4
grpcio==1.35.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.1
pep517==0.8.2
Pillow==8.1.0
progress==1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dotenv==0.15.0
python-http-client==3.3.1
pytoml==0.1.21
pytz==2020.5
regex==2020.11.13
requests==2.22.0
retrying==1.3.3
rsa==4.7
six==1.14.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.25.8
webencodings==0.5.1
djangocms-redirect==0.6.0
psycopg2-binary==2.8.6